Toolset directory reorganization
