+++ toolset/benchmark/README.md
@@ -1,0 +1,8 @@
+# Toolset Benchmark Files
+
+This directory contains files that implement or support the benchmark feature:
+
+* `benchmarker.py`: Benchmark execution methods. Called by `run-tests.py`.
+* `framework_test.py`: Methods to run a framework's tests. Called by `benchmarker.py`.
+
+These files are not meant for you to run directly. For instructions on running the benchmark, please refer to the [parent directory's README file](../README.md).
