+++ toolset/setup/linux/__init__.py
@@ -1,0 +1,1 @@
+# Linux server and client setup
